🏗️ refactor: breaking project to smaller components
breaking global nav component to new smaller global nav component and global nav item component
