--- ui/global-nav.tsx
@@ -1,98 +1,0 @@
-'use client';
-
-import { items, type Item } from '#/lib/items';
-import { CapsLogo } from '#/ui/caps-logo';
-import { MenuAlt2Icon, XIcon } from '@heroicons/react/solid';
-import clsx from 'clsx';
-import Link from 'next/link';
-import { useSelectedLayoutSegment } from 'next/navigation';
-import { useState } from 'react';
-
-export function GlobalNav() {
-  const [isOpen, setIsOpen] = useState(false);
-  const close = () => setIsOpen(false);
-
-  return (
-    <div className='fixed top-0 z-10 flex w-full flex-col border-b border-gray-800 bg-black lg:bottom-0 lg:z-auto lg:w-72 lg:border-b-0 lg:border-r lg:border-gray-800'>
-      <div className='flex h-14 items-center px-4 py-4 lg:h-auto'>
-        <Link
-          href='/'
-          className='group flex w-full items-center gap-x-2.5'
-          onClick={close}
-        >
-          <div className='h-10 w-10 rounded-full border border-white/20 p-2 group-hover:border-white/60'>
-            <CapsLogo />
-          </div>
-
-          <h3 className='font-semibold tracking-wide text-gray-400 group-hover:text-gray-50'>
-            Capsmark
-          </h3>
-        </Link>
-      </div>
-
-      <button
-        type='button'
-        className='group absolute right-0 top-0 flex h-14 items-center gap-x-2 px-4 lg:hidden'
-        onClick={() => setIsOpen(!isOpen)}
-      >
-        <div className='font-medium text-gray-100 group-hover:text-gray-400'>Menu</div>
-        {isOpen ? (
-          <XIcon className='block w-6 text-gray-400' />
-        ) : (
-          <MenuAlt2Icon className='block w-6 text-gray-400' />
-        )}
-      </button>
-
-      <div
-        className={clsx('overflow-y-auto lg:static lg:block', {
-          'fixed inset-x-0 bottom-0 top-14 mt-px bg-black': isOpen,
-          hidden: !isOpen,
-        })}
-      >
-        <nav className='space-y-6 px-2 pb-24 pt-5'>
-          {items.map(({name, items}) => {
-            return (
-              <div key={name}>
-                <div className='mb-2 px-3 text-xs font-semibold uppercase tracking-wider text-gray-400/80'>
-                  <div>{name}</div>
-                </div>
-
-                <div className='space-y-1'>
-                  {items.map((item, index) => (
-                    // todo add anchor link here
-                    <GlobalNavItem key={index} item={item} close={close} />
-                  ))}
-                </div>
-              </div>
-            );
-          })}
-        </nav>
-      </div>
-    </div>
-  );
-}
-
-function GlobalNavItem({ item, close }: { item: Item; close: () => false | void }) {
-  // can be usefull to detect the # tag
-  const segment = useSelectedLayoutSegment();
-  
-  // todo add logic to check if is active
-  const isActive = true;
-
-  return (
-    <Link
-      onClick={close}
-      href={`/`}
-      // href={`/${item._id_can_be_added_to_object}`}
-      className={clsx(
-        'block rounded-md px-3 py-2 text-sm font-medium hover:text-gray-300',
-        {
-          'text-gray-400 hover:bg-gray-800': !isActive,
-          'text-white': isActive,
-        },
-      )}
-    >
-      {item.name}
-    </Link>
-  );
-}
